requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.2
djoser==2.2.2
django-cors-headers==4.3.1
drf-spectacular==0.26.5
Pillow==10.1.0
python-decouple==3.8
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
